+++ CHANGELOG.md
@@ -1,2 +1,14 @@
-# Change Log
+# Changelog
 This project adheres to [Semantic Versioning](http://semver.org/).
+
+## [0.1.0] - 2018-05-25
+### Added
+- Added the option to break multiple selectors to new lines
+
+### Fixed
+- Selectors (after the first) within blocks were not respecting their initial indentation
+
+
+## [0.0.1] - 2018-05-24
+### Added
+- Added the initial spacing functionality for this plugin